+++ guizhanlib-common/src/main/java/net/guizhanss/guizhanlib/common/utils/JsonUtil.java
@@ -1,15 +1,12 @@
-package net.guizhanss.guizhanlib.utils;
+package net.guizhanss.guizhanlib.common.utils;
 
 import com.google.common.base.Preconditions;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
-import com.google.gson.stream.JsonReader;
 import lombok.experimental.UtilityClass;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.io.BufferedReader;
 import java.util.regex.Pattern;
 
 /**
@@ -20,43 +17,8 @@ import java.util.regex.Pattern;
 @UtilityClass
 @SuppressWarnings("ConstantConditions")
 public final class JsonUtil {
+
     private static final Pattern DOT_SPLIT = Pattern.compile("\\.");
-
-    /**
-     * Parse json {@link String}.
-     *
-     * @param json The string to be parsed.
-     * @return The parsed {@link JsonElement}.
-     */
-    @Nonnull
-    public static JsonElement parse(@Nonnull String json) {
-        Preconditions.checkArgument(json != null, "JSON string should not be null");
-        return new JsonParser().parse(json);
-    }
-
-    /**
-     * Parse json from {@link BufferedReader}.
-     *
-     * @param reader The {@link BufferedReader} to read json from.
-     * @return The parsed {@link JsonElement}.
-     */
-    @Nonnull
-    public static JsonElement parse(@Nonnull BufferedReader reader) {
-        Preconditions.checkArgument(reader != null, "reader should not be null");
-        return new JsonParser().parse(reader);
-    }
-
-    /**
-     * Parse json from {@link JsonReader}.
-     *
-     * @param reader The {@link JsonReader} to read json from.
-     * @return The parsed {@link JsonElement}.
-     */
-    @Nonnull
-    public static JsonElement parse(@Nonnull JsonReader reader) {
-        Preconditions.checkArgument(reader != null, "reader should not be null");
-        return new JsonParser().parse(reader);
-    }
 
     /**
      * The value from given path of {@link JsonObject}.