+++ guizhanlib-common/src/main/java/net/guizhanss/guizhanlib/common/utils/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package contains utility methods.
  */
-package net.guizhanss.guizhanlib.utils;
+package net.guizhanss.guizhanlib.common.utils;