+++ guizhanlib-common/src/main/java/net/guizhanss/guizhanlib/java/package-info.java
@@ -1,4 +1,0 @@
-/**
- * 该包包含某些Java类型的相关助手
- */
-package net.guizhanss.guizhanlib.java;